+++ src/plone/restapi/services/workflow/info.py
@@ -1,13 +1,31 @@
 # -*- coding: utf-8 -*-
 from Products.CMFCore.utils import getToolByName
+from Products.CMFCore.interfaces._content import IWorkflowAware
+from plone.restapi.interfaces import IExpandableElement
 from plone.restapi.serializer.converters import json_compatible
 from plone.restapi.services import Service
+from zope.component import adapter
+from zope.interface import Interface
+from zope.interface import implementer
 
 
-class WorkflowInfo(Service):
-    """Get workflow information
-    """
-    def reply(self):
+@implementer(IExpandableElement)
+@adapter(IWorkflowAware, Interface)
+class WorkflowInfo(object):
+
+    def __init__(self, context, request):
+        self.context = context
+        self.request = request
+
+    def __call__(self, expand=False):
+        result = {
+            'workflow': {
+                "@id": '{}/@workflow'.format(self.context.absolute_url()),
+            },
+        }
+        if not expand:
+            return result
+
         wftool = getToolByName(self.context, 'portal_workflow')
         history = wftool.getInfoFor(self.context, "review_history")
 
@@ -29,7 +47,16 @@ class WorkflowInfo(Service):
                     action['review_state'],
                     self.context.portal_type))
 
-        return {
+        result['workflow'].update({
             'history': json_compatible(history),
             'transitions': transitions,
-        }
+            })
+        return result
+
+
+class WorkflowInfoService(Service):
+    """Get workflow information
+    """
+    def reply(self):
+        info = WorkflowInfo(self.context, self.request)
+        return info(expand=True)['workflow']